Fixed location detection on Android 15

+++ lib/model/weather.dart
@@ -11,7 +11,7 @@ class Weather {
   factory Weather.fromJson(Map<String, dynamic> json) {
     return Weather(
       cityName: json['name'],
-      temperature: json['main']['temp'],
+      temperature: json['main']['temp'].toDouble(),
       mainCondition: json['weather'][0]['main'],
     );
   }
